MB-4234 Move update agent validator code to new files
