+++ pkg/services/mto_agent/mto_agent_validators.go
@@ -1,0 +1,167 @@
+package mtoagent
+
+import (
+	"fmt"
+
+	"github.com/getlantern/deepcopy"
+	"github.com/gobuffalo/validate"
+	"github.com/gofrs/uuid"
+
+	"github.com/transcom/mymove/pkg/models"
+	"github.com/transcom/mymove/pkg/services"
+)
+
+// UpdateMTOAgentBaseValidator is the key for generic validation on the MTO Agent
+const UpdateMTOAgentBaseValidator string = "UpdateMTOAgentBaseValidator"
+
+// UpdateMTOAgentPrimeValidator is the key for validating the MTO Agent for the Prime contractor
+const UpdateMTOAgentPrimeValidator string = "UpdateMTOAgentPrimeValidator"
+
+// UpdateMTOAgentValidators is the map connecting the constant keys to the correct validator
+var UpdateMTOAgentValidators = map[string]updateMTOAgentValidator{
+	UpdateMTOAgentBaseValidator:  new(baseUpdateMTOAgentValidator),
+	UpdateMTOAgentPrimeValidator: new(primeUpdateMTOAgentValidator),
+}
+
+type updateMTOAgentValidator interface {
+	validate(agentData *updateMTOAgentData) error
+}
+
+// baseUpdateMTOAgentValidator is the type for validation that should happen no matter who uses this service object
+type baseUpdateMTOAgentValidator struct{}
+
+func (v *baseUpdateMTOAgentValidator) validate(agentData *updateMTOAgentData) error {
+	err := agentData.checkShipmentID()
+	if err != nil {
+		return err
+	}
+
+	err = agentData.getVerrs()
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// primeUpdateMTOAgentValidator is the type for validation that is just for updates from the Prime contractor
+type primeUpdateMTOAgentValidator struct{}
+
+func (v *primeUpdateMTOAgentValidator) validate(agentData *updateMTOAgentData) error {
+	err := agentData.checkShipmentID()
+	if err != nil {
+		return err
+	}
+
+	err = agentData.checkPrimeAvailability()
+	if err != nil {
+		return err
+	}
+
+	err = agentData.checkContactInfo()
+	if err != nil {
+		return err
+	}
+
+	err = agentData.getVerrs()
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// updateMTOAgentData represents the data needed to validate an update on an MTOAgent
+type updateMTOAgentData struct {
+	updatedAgent        models.MTOAgent
+	oldAgent            models.MTOAgent
+	availabilityChecker services.MoveTaskOrderChecker
+	verrs               *validate.Errors
+}
+
+// checkShipmentID checks that the user didn't attempt to change the agent's shipment ID
+func (v *updateMTOAgentData) checkShipmentID() error {
+	if v.updatedAgent.MTOShipmentID != uuid.Nil && v.updatedAgent.MTOShipmentID != v.oldAgent.MTOShipmentID {
+		v.verrs.Add("mtoShipmentID", "cannot be updated")
+	}
+
+	return nil
+}
+
+// checkPrimeAvailability checks that agent is connected to a Prime-available shipment
+func (v *updateMTOAgentData) checkPrimeAvailability() error {
+	isAvailable, err := v.availabilityChecker.MTOAvailableToPrime(v.oldAgent.MTOShipment.MoveTaskOrderID)
+
+	if !isAvailable || err != nil {
+		return services.NewNotFoundError(v.updatedAgent.ID, "while looking for Prime-available MTOAgent")
+	}
+
+	return nil
+}
+
+// checkContactInfo checks that the new agent has the minimum required contact info: First Name and one of Email or Phone
+func (v *updateMTOAgentData) checkContactInfo() error {
+	firstName := v.oldAgent.FirstName
+	if v.updatedAgent.FirstName != nil {
+		firstName = v.updatedAgent.FirstName
+	}
+
+	// Check that we have something in the FirstName field:
+	if *firstName == "" || firstName == nil {
+		v.verrs.Add("firstName", "cannot be blank")
+	}
+
+	email := v.oldAgent.Email
+	if v.updatedAgent.Email != nil {
+		email = v.updatedAgent.Email
+	}
+
+	phone := v.oldAgent.Phone
+	if v.updatedAgent.Phone != nil {
+		phone = v.updatedAgent.Phone
+	}
+
+	// Check that we have one method of contacting the agent:
+	if (*email == "" || email == nil) && (*phone == "" || phone == nil) {
+		v.verrs.Add("contactInfo", "agent must have at least one contact method provided")
+	}
+
+	return nil
+}
+
+// getVerrs looks for any validation errors and returns a formatted InvalidInputError if any are found.
+// Should only be called after the other check methods have been called.
+func (v *updateMTOAgentData) getVerrs() error {
+	if v.verrs.HasAny() {
+		return services.NewInvalidInputError(v.updatedAgent.ID, nil, v.verrs, "Invalid input found while validating the agent.")
+	}
+
+	return nil
+}
+
+// setNewMTOAgent compares updatedAgent and oldAgent and updates a new MTOAgent instance with all data
+// (changed and unchanged) filled in
+func (v *updateMTOAgentData) setNewMTOAgent(newAgent *models.MTOAgent) error {
+	err := deepcopy.Copy(&newAgent, &v.oldAgent)
+	if err != nil {
+		return fmt.Errorf("error copying agent data %w", err)
+	}
+
+	if v.updatedAgent.MTOAgentType != "" {
+		newAgent.MTOAgentType = v.updatedAgent.MTOAgentType
+	}
+	if v.updatedAgent.FirstName != nil {
+		newAgent.FirstName = v.updatedAgent.FirstName
+	}
+	if v.updatedAgent.LastName != nil {
+		newAgent.LastName = v.updatedAgent.LastName
+	}
+	if v.updatedAgent.Email != nil {
+		newAgent.Email = v.updatedAgent.Email
+	}
+	if v.updatedAgent.Phone != nil {
+		newAgent.Phone = v.updatedAgent.Phone
+	}
+
+	return nil
+}
